core, restapi: provide capabillity to get vm ovf config
This patch adds the capabllity to get the current vm configuration (if
possible) through the ovirt rest-api.

The usage is done through a GET request to /api/vms or /api/vms/GUID/
with All-Content header with 'true' as the value.

If the VM OVF configuration couldn't be generated succesfully (if the vm
is currentky locked for example), the configuration won't be returned as
part of the result.

Change-Id: I5cbcd666519ffd138fd3efd4e130495529eaec0f
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1047965
Signed-off-by: Liron Aravot <laravot@redhat.com>