engine: Modify EnableMACAntiSpoofing version to general
There used to be a mix in the declaration and usage of
MacAntiSpoofingFilterRulesSupported configuration value
which allowed the user to determine in which cluster
level the feature should be used. This should be determined
by the engine itself and not by the user.

Therefore the versioning is removed from this parameter
which will define for the entire system if no-mac-spoofing
filter should be applied to running vms or not.

Change-Id: Ia2c7659844318dd38241a6ae56895ca3c4bf91cf
Bug-Url: https://bugzilla.redhat.com/1063711
Signed-off-by: Moti Asayag <masayag@redhat.com>