tools: notifier: fix NotificationPropertiesTest.
Call NotificationProperties.release() so we do not
get a stale copy of Notification Properties.

(Set default and get instance are not enough
if the current class loader has a copy of
NotificationProperties with an old instance
we will get that one!)

also enable NotificationPropertiesTest to run both under maven and straight
from source thus speeding up testing during development.

Change-Id: Ia3488856323577e3db6c779d86c190e064e15c97
Signed-off-by: Mooli Tayer <mtayer@redhat.com>