restapi: Add ability to extract rsdl.xml from jar
A previous patch moved RSDL generation to build time.
This patch adds a class with a main() method which
copies rsdl.xml, rsdl_gluster.xml and api.xsd to the
desired directory.

The idea behind this, is that sdk generation will be
done completely automatically by maven goals. To copy
the files mentioned above (to a directory named, for
example, '/service') - add the following configuration
to pom.xml of any project that depends on
restapi-definition:

    <profile>
      <id>build-rsdl</id>
      <activation>
        <activeByDefault>true</activeByDefault>
      </activation>
      <build>
        <plugins>
          <plugin>
            <groupId>org.codehaus.mojo</groupId>
            <artifactId>exec-maven-plugin</artifactId>
            <executions>
              <execution>
                <phase>compile</phase>
                <goals>
                  <goal>java</goal>
                </goals>
                <configuration>
                  <mainClass>org.ovirt.engine.api.rsdl.RsdlIOManager</mainClass>
                  <arguments>
                    <argument>/service</argument>
                  </arguments>
                </configuration>
              </execution>
            </executions>
          </plugin>
        </plugins>
      </build>
     </profile>

Change-Id: I11834dac081e17c702c614ae923575a3debba6eb
Signed-off-by: Ori Liel <oliel@redhat.com>