core: modify log message on live snapshot failure
The log message that was printed when live snapshot failed contained the
word "Wasn't" which appears in the log file as "Wasnt". That is
confusing because if you use copy-paste in order to look for it in the
log file you won't find it. Thus, the log message is changed.

Change-Id: Iee77b46be0948ab90de2e9ea438ef9e9b4413546
Signed-off-by: Arik Hadas <ahadas@redhat.com>