engine: Refactor AuditLog severity handling - step 1
Having a separate structure (AuditLogDirector.severities) to hold
the severity definitions for messages from AuditLogType has so far
caused many warnings and bug reports. Developers were forgetting
to update the map.

Java supports fields in Enums and we were already using them to
limit the frequency of reporting for certain messages.

This patch moves the severity to the AuditLogType enum. NORMAL
severity is considered default and does not have to be specified.

It keeps the severities map and adds a test that compares the old
map with the new field to make sure no mistake was made during the
rewrite.

Change-Id: I9c784a2f63d1c1d194f8f6712a39fa18a21b17da
Signed-off-by: Martin Sivak <msivak@redhat.com>