tools: notifier: implement and use transport.idle().
As part of transitioning to more then one Transport,
It is no longer possible to let a transport sleep
and retry for as long as it wants. that might
starve other transports especially in a short
notification interval.

Retrying transports (those that can detect sending failures)
Should implement a queue and attempt to send it on idle.

This patch also fixes a bug where smtp loging is done
only before the first send attempt.

Change-Id: I8c4bafb542d28cb584e0751446d3e327f93e8112
Related-To: https://bugzilla.redhat.com/show_bug.cgi?id=1051492
Signed-off-by: Mooli Tayer <mtayer@redhat.com>
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>