engine: Reorder network actions when changing host cluster
When moving host between clusters, there is a need to first
clear all the labeled networks originated in the source
cluster. Second, to add all of the labeled networks according
to the target cluster attachment.

The reason for it is to deallocate assigned interfaces which
were assigned with networks of the source cluster.

Change-Id: I1ea644edbe7e3d314d4d1baf8a45b52d7a5a8f95
Bug-Url: https://bugzilla.redhat.com/1064953
Signed-off-by: Moti Asayag <masayag@redhat.com>