core: Fixing I18N validation to allow empty strings
The I18N validation in the validation framework was set to pass validation
an empty string, this behavior is incorrect since the I18N validation should
only care for extisting characters within the string and since an empty
string contains no illegal characters, validation should pass, for
validating a strings length the size validation should be used.
The regular expression was changed to pass validation on an empty string
and the places which used this validation without minimum string length
validation were changed to validate a minimum string length of 1 in order
not to have regressions.
This solves a validation problem when creating a VM disk via REST and
not specifying an alias, validation should not fail in that case and an
alias should be generated by the add disk command

Signed-off-by: Tal Nisan <tnisan@redhat.com>
https://bugzilla.redhat.com/show_bug.cgi?id=1065727
Change-Id: I6d6cf47a0ccaf1d8af3d46e5dee9c264e9a29bb3