core: remove redundant transaction on run vm
Saving the VM dynamic information that we get back from VDSM after
calling VDSM to create a VM was used to be executed in transaction
because we used to update information related to the VDS in the DB
and we wanted it to be atomically.

Now that we don't update anything else in the DB besides the dynamic
information of the VM, doing the update in transaction with Required
scope is redundant and we can drop the transaction.

Change-Id: I9819a9defe421deab3badf1e4e1e62e7e3e3e92f
Signed-off-by: Arik Hadas <ahadas@redhat.com>