core: change asyncRunningVms to be Set
ResourceManager#_asyncRunningVms used to be of type ConcurrentHashMap
but provided functionality of Set. Since Java 1.6 there is an easy
way to create thread-safe Set which is based on ConcurrentHashMap. By
setting the asyncRunningVms collection to such Set, our code becomes
more expressive and easier to read, and we don't change the correctness
since we are using the same underlying implementation (one difference
though is that the implementation now uses put instead of putIfAbsent,
but it does not matter because we are always using the same value).

This patch also contains minor refactoring in ResourceManager.

Change-Id: I807936d76ee7298a126e7dd61d75cbb345e06829
Signed-off-by: Arik Hadas <ahadas@redhat.com>