core: set DC compatability version to latest...
Set DC compatibility version to latest when null

Up to now when adding a DC from API without giving explicitly the
compatibility version was resulted in a failure on unsupported version.

This patch changes this behaviour by setting the latest version as a
default if a null was given for compatibility version.

Change-Id: I53373e9ce2563d91a96353f068cd19069508399d
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1053062
Signed-off-by: Eli Mesika <emesika@redhat.com>