core: GetAllFromVmStaticByStoragePoolId - remove unneeded join
remove unneeded join with vm_dynamic, vm_dynamic and vm_static should be
added in the same transacation when adding a vm. I assume that this join
was added to avoid querying templates, but it's unneeded as we already
check the entity type in the query without performing that probably more
expensive operation.

Change-Id: I4fd372a5c4018d3c7bb7e38e1375fa198a14b572
Signed-off-by: Liron Aravot <laravot@redhat.com>