core: resouces of VMs in launch state shouldn't be committed
Currently we count pending in scheduling process, and committed memory
is calculated in server side according to VMs attached to the host
(including VMs in wait for launch state- host in VM entity
is update when CreateVm command returns from VDSM).
Pending memory is being release in powering-up state, so in fact
we count resources twice: since the time VM is in wait for launch till it's
powering up.

Memory commitment in VDSM occurs after a vm was requested, so using
this data will eliminate VM throttling that failed to reach powering up
state.

Change-Id: Ie2db20c213ac326d620fad578bb5c1c78ff0b3bc
Bug-Url: https://bugzilla.redhat.com/1074478
Signed-off-by: Gilad Chaplik <gchaplik@redhat.com>