fingbugs: Ignore BC_BAD_CAST_TO_ABSTRACT_COLLECTION in uicommonweb module
Ignore all the questionable casts from Collection to abstract collection in this module.
Unfortunately, these warnings are currently inherit to the design, ListModel base items are
of a type Collection and many times need to be cast to a List for sorting purposes etc.
Handling those on an individual basis just adds noise to the process.

Change-Id: I4e8811fd1fae838c333866627e94699fa73a3e15
Signed-off-by: Tal Nisan <tnisan@redhat.com>