core: added filtering to ensure only the templates return
If the search type in search query was "VmTemplate" it anyway returned all the
content of the vm_templastes_view which is a union of instance types, images
and templates.

Fixed by adding the filtering to the SearchQuery directly and not the DB level
because the vm_templates_view is used in a way that it expects that all three
types are present in it.

Change-Id: Id2b8366f9099fd4461ae7f876fda782e0c8568ab
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>