core: extract parameters initialization from vdsUpEvent
Extract the part that initialize the parameters for the migrate VM to
server commands from VdsEventListener#vdsUpEvent to make it more
readable.

In addition, change MigrateBrokerVDSCommand to create the migrate-info
map right before invoking VDSM verb as this information is not needed
until then and we do not need to hold it as class member.

Minor cleanup in MigrateVDSCommand is also included in this patch.

Change-Id: I54501c64f9b0ff3cd2bb80d457a597d01328b8d5
Signed-off-by: Arik Hadas <ahadas@redhat.com>