core: make migrate commands non-transactive
Patch 94e48a02 removed the global transaction during RunVmCommand.
Global transactions should also be removed from the migrate commands
for the same reasons and because it prevents us from migrating multiple
VMs from different hosts to the same host while one of the migrations
is in its execute phase or start a VM on a host while another VM is
migrating to that host and the migration is in its execute phase.

Change-Id: I4894ebd5f062174f9780926a0ce6b499fcc0fc7b
Bug-Url: https://bugzilla.redhat.com/1064931
Signed-off-by: Arik Hadas <ahadas@redhat.com>