core: detach/remove of domain with ovf stores
*When detaching a storage domain the engine should allow to detach it
even if it has disks that are used as the domain ovf store.

*When removing a storage domain, use the StorageDomainDao.remove method,
this method is currently being used by the ForceRemoveStorage flow and
cleans up anything associated with the domain.
As we allow to remove a domain with ovf stores, the domain disks should
be removed as well (handled by this sp).
This sp is heavier than the previous calls, but we are saving round
trips and removing storage domain is a rare scenario so we better
make sure that no leftovers are left by executing this sp rather than
take care of that ourselves and rely on bll CDA checks.

Change-Id: Icde56b54bd11c617056c819fcdbacac71753e264
Signed-off-by: Liron Aravot <laravot@redhat.com>