frontend: Enable configuration of Serial Number
Enable configuration of serial number policies for Vm, Template and
Cluster.

The editor functionality is encapsulated in SerialNumberPolicyWidget
that edits the SerialNumberPolicyModel which is embedded in the
UnitVmModel and ClusterModel.

The SerialNumberPolicyWidget utilizes the existing EnumRadioEditor to
edit the SerialNumberPolicy. The EnumRadioEditor was moved from webadmin
to gwt-common to be also usable when editing VM's in userportal.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=918138
Change-Id: I7064570c2d66fdd473c31214892867b8b9b67a21
Signed-off-by: Martin Betak <mbetak@redhat.com>