core: Preserve bonding options when updating vlan over bond
The ovirt-engine used to erase any pre-configured
bonding options, if it wasn't provided explicitly
by the api users.

In order to preserve the bonding options properly,
there is a need either to examine it from the supplied
properties or to use the already configured bonding
device from the engine db.

Change-Id: I09b150b11913a9d8cddc5115e28c027e7c944910
Bug-Url: https://bugzilla.redhat.com/1068970
Signed-off-by: Moti Asayag <masayag@redhat.com>