webadmin: Remove API version field for Neutron
Only API version 2.0 is curently supported, and since the user has a
prefilled value in the URL field suggesting him the correct server URL
to put, there is no need for him to be able to "select" the version.

In the future should any other API version be available/supported, it
should be solved in an appropriate manner (i.e. auto negotiate the
version).

Change-Id: Ic264c97b4acb82ffee298cee3f89340cda18ddc3
Bug-Url: https://bugzilla.redhat.com/987917
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>