core: genericSearch method which does not get filter
We used to have SearchQuery#genericSearch method that gets filter as
a parameter and in case we didn't want to filter, we set the filter
to null. A more standard way to do it is by adding a method that
does not get filter and delegates to the method that gets a filter
by passing null as the filter.

Change-Id: Ib013b8eb2baa77b6adc01d70c3df0a87482dc296
Signed-off-by: Arik Hadas <ahadas@redhat.com>