webadmin: block reinitialize on uninitalized pool
When a pool is uninitalized there's no need to allow reinitializing
it, any wanted domain can be attached to the pool.
Allowing to perform reinitalize in that status caused to NPE in the
engine which currently attempts to load the current master domain of the
pool.
As there's no meaning for recovery/reinitialize in that scenario, it can
be blocked - as this issue was never encountered, we can add an engine
verification if and when it will be necessary.

Change-Id: I480b1f99ae881d43d34342ccb3bb248e60f556ce
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1069581
Signed-off-by: Liron Aravot <laravot@redhat.com>