webadmin: Adding Quota column to the Disk tab
Adding a new column to the Disk tab, Quota.
the new Quota column will be visible only in case
the user point to a DC in the tree view,
and that DC is not marked as Quota disabled.
The possible value for this column is:
Empty(null) - no Quota was selected for this Disk.
String value - the name of the Quota selected for this Disk.

Change-Id: I2332da722396a16aca9545b9ef0532ebc84d8d5e
Bug-Url: https://bugzilla.redhat.com/949863
Signed-off-by: Kobi Ianko <kobi@redhat.com>