engine: added proper instance type handling for the watchdog
1:
Added support for adding the watchdog also for the UpdateTemplateCommand.

2:
If the watchdog is added/updated on the instance type, the specific validation
should not be run since you can add any device to the instance type.
The proper validations are added only when creating the specific VM from the
instance type.

Change-Id: Ibd953828b7bc3ce1607377a7db28970f5ab7a03e
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>