restapi: hosted engine maintenance operations
Add REST API information/commands for hosted engine maintenance.

In the host XML, a <HostedEngine> tag will provide information about
Hosted Engine on that host (active, configured, score, global and
local maintenance status).  This will be present only if the request
contains the 'All-Content: true' and the HA agent is configurd on
the host.

Additionally, a new VM action, hosted_engine_maintenance, is provided
allowing global maintenance mode to be enabled/disabled through the
host the engine vm is currently running on.

Change-Id: I72bb087eb6037c2c9f039e989b7261b54534adee
Bug-Url: https://bugzilla.redhat.com/1073051
Signed-off-by: Greg Padgett <gpadgett@redhat.com>