core: don't load disks for templates on search
When searching for templates,
the search loads all the disks for each template.
this is not necessary as disks are a different collection that need to
be fetched upon request,
also the disks are not used currently by the search initiators,
and it make the templates search slow for systems with many templates.

this patch removes the disks population on templates search.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1070265
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1064907
Change-Id: Ia83a3117f2c65bbb17972095cf3271eb39449cd6
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>