webadmin: [Fix] Template changes do not update OS
This change forces an OS list update in the "New Pool" and the
"Edit Pool" dialogs after the user selects a template. It was
introduced in order to avoid a class of bugs where the OS list
could show OSes for another architecture during the creation or
edition of VM Pools.

Change-Id: I959cdd6d133ead4f83cfb6cd68145519efb25489
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1067735
Signed-off-by: Leonardo Bianconi <leonardo.bianconi@eldorado.org.br>
Signed-off-by: Gustavo Pedrosa <gustavo.pedrosa@eldorado.org.br>