engine: Collect stats based on non-operational reason
If a host become non-operational due to network
issue, the statistics will not be collected any
more. Therefore any attempt to activate the host
by the Auto-Recovery manager will be skipped.
Even if the network issue will be solved, the host
will remain non-operational unless activated by
the user.

This patch will allow to examine the cause for the non-
operational status and continue stats collection to
assure a network issue was solved and re-attempt to
activate the host.

Change-Id: I407d165e1cf13bbd1ba50811eaed4771c3f36f1a
Bug-Url: https://bugzilla.redhat.com/1070260
Signed-off-by: Moti Asayag <masayag@redhat.com>