engine: Set boot order to more then one CDROM if selected
When dealing with more then one CDROM we set only the first in list if
user select boot order from CDROM, which can cause a problem if the
first CDROM is payload or cloud-init which are not bootable.
When setting all CDROM to be bootable qemu will try to boot from one
of them.

Change-Id: I7d4783f084623bc59096d87514a6f143e056d5a2
Bug-Url: https://bugzilla.redhat.com/1064927
Signed-off-by: Shahar Havivi <shaharh@redhat.com>