engine: Set boot order to more then one CDROM if selected
When dealing with more then one CDROM we set only the first in list if
user select boot order from CDROM, which can cause a problem if the
first CDROM is payload or cloud-init which are not bootable.
When setting all CDROM to be bootable qemu will try to boot from one
of them.

Change-Id: I1f0b8b6a8346dd71b716df8d90f852de550bae97
Bug-Url: https://bugzilla.redhat.com/1064927
Signed-off-by: Shahar Havivi <shaharh@redhat.com>