webadmin: Fix Default OS selection for PPC
The change 23809 introduced a problem with the default OS for PPC64,
since the id 0 was fixed in the AsyncDataProvider class. This change
was supposed to fix a possible issue, that happens when the OS id added
to a VM is missing in the OSInfo property file. I reverted the solution
that was using the class MapWithDafults, because there are two default
OSes (one for each architecture), not only one.
I couldn't figure out how to fix this issue, maybe when starting the
system and exit with an error or show error on event log. It deserves
more discussion.

Change-Id: I02843c1af1b3aecf2d3966bf836395550b0be554
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1062344
Signed-off-by: Leonardo Bianconi <leonardo.bianconi@eldorado.org.br>