aaa: Fix group population
The following patch fixes group population for users and groups -
a. It does not use DbFacade to get the external id - external id
should be retrieved from the directory
b. It runs the group population in all cases user information is populated
c. At group population - the recrusion is fixed/changed - the LdapGroup
object that is created is at ProceedGroupSearchResult , before making
the recursive calls at populateGroup.

Topic: AAA
Change-Id: I599fca005b27f6873cda3e5a46081de3357089b2
Signed-off-by: Yair Zaslavsky <yzaslavs@redhat.com>