core: change vm custom properties validation
Change the validation of custom properties of VM when we run it,
such that it will check also the custom properties that resides
within the parameters of RunVm (which are set in run once mode).

This change will allow us to set the custom properties which
resides in the VM object on RunVmOnceCommand from the parameters
after the can do action phase.

In addition, minor refactoring was made to the validation of boot
sequence on run VM.

Change-Id: I66cec8aa76958831924a48b8a7805cf0d181dcf2
Signed-off-by: Arik Hadas <ahadas@redhat.com>