core: Fix group id on updating cache of users
When db users cache is updated in refreshAllUsers, also user groups
membership is updated. If some group is not in local db, by mistake
externald id of group is set to user groups instead of internal one.
This patch sets empty GUID for groups that are not stored in local db.

Change-Id: I84d04420bc508f482c6ed4466e6937d8741bfe62
Bug-Url: https://bugzilla.redhat.com/1073235
Signed-off-by: Martin Perina <mperina@redhat.com>