engine: Remove port if hot plug new ext. nic fails
If a new vNIC is added on an external network, and this vNIC is also hot
plugged as part of this flow then a corresponding entity will be created
on the external provider.
If the aforementioned action fails, the vNIC will be removed from the DB
(by compensation mechanism) but not from the external provider, leaving
garbage forever on it.

In this fix, a new vNIC that fails to hot plug will also be removed from
the external provider.
At this point, there is no need to remove an existing vNIC
should the hot plug fail, as the entity that was created on the external
provider should be removed only when the vNIC is being removed, or the
network is not used by the vNIC anymore.

Change-Id: Iddeac63a47286e9b83ac02fb2e55f7dd56092805
Bug-Url: https://bugzilla.redhat.com/987814
Signed-off-by: Mike Kolesnik <mkolesni@redhat.com>
(cherry picked from commit c3fcf1485b5b15e4b9509b910b24202f39ea594c)