webadmin: host right click menu is missing 'reinstall' option
Host's right click menu is missing "reinstall" option.
Even when it's in maintenance, there is no such option.
Instead, there's a link at the bottom left part of host's
General sub tab - which is not likely the place where user will
search for this option since the rest of this tab
is for info about the host (read only).

Removes the install link and message from the General
Sub Tab.

Change-Id: Ib716a984904d6b299cefa981d6d36d5e257e3f1c
Bug-Url: https://bugzilla.redhat.com/1065739
Signed-off-by: Ravi Nori <rnori@redhat.com>