core: change domains status to unknown when there are no reporting hosts
Currently the monitored pool domains 'change' their statuses according to
the reports received by the pool hosts that are in status UP.

When there are no reporting hosts - the status of the pool domains that are
monitored should be changed to unknown, because otherwise nothing will trigger
the status change. In that case (an example) we might be left with a domain
in 'active' status although there are no reporting hosts.

Change-Id: I8091a4864711aeccee41effb1bc7d9823a1870c7
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=975742
Signed-off-by: Liron Aravot <laravot@redhat.com>