engine: Networks with role configured with DHCP in nic
Network with roles must have an IP address in order for
the host to be operational (either for display or for
migration purposes).

A mass operation to configure networks on various hosts,
where some of the networks might have roles, is by setting
the boot protocol to DHCP. This will configure the networks
on the hosts in a way that will enable to IP assignment.

If the DHCP request fails, there is a change the network
will not be configure on the host. This will cause the host
to be non-operational, with a simple resolution for the
nature of the problem.

Change-Id: I3ee53ec1f354ea766cea59977e869f97c2be94df
Bug-Url: https://bugzilla.redhat.com/1073043
Signed-off-by: Moti Asayag <masayag@redhat.com>