core: Change Vm status on destroy only for SRC VMs
when stopping a migrating VMs, do the state change only if the VM is
really running on SRC host.

StopVm shoots 2 destroy commands incase the vm is migrating - 1 for the
DST and one for the SRC. after the DST is destroy the Vm must not be
DOWN, otherwise an interleaving RunVm could start
it while we're about to shoot the next stop.

that solution is reverting change
I6109ef0ea4869b59a93be6fef4b0c3cff50da62e
which implemented that with an engine lock

Change-Id: I422e75fc863e3416261a4d3eaef4312dc933bdac
Bug-url: https://bugzilla.redhat.com/1064859
Signed-off-by: Roy Golan <rgolan@redhat.com>