backend: Set Vm.current_cd on RunOnce
Modified tracking of current CDs.
Before the persistent configuration was stored in VmBase#isoPath and if
we got another "runtime" value from VDSM we would store that in
VmDynamic#currentCd. Otherwise currentCd would be null and when getting
the "current" status one would first have to check if currentCd is null
and if not then fall back to isoPath.

This was not only difficult to use from other parts of code, but it also
didn't work for RunOnce where the VDSM would not report the 'cdrom' in
getAllVmStats.

Now we set the currentCd directly in RunVmCommand and can always access
directly the "current" value of cd rom.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1065719
Change-Id: I5bfc4fbfae15e902898859dddf3cdffe9301dab2
Signed-off-by: Martin Betak <mbetak@redhat.com>