core: VURTI - fix handle expection in GetVmStats
after getting all vms with List command, VURTI loops through them
to update internal structure data. BUT if the call fails, VURTI removes
the VM from the internal running map - this leads to a split-brains
for VURTI assumes the Vm is DOWN and if it HA, its runs it.

the fix is to exclude this Vm from the list of the vms that have changed
this is similar to a case where Vm didn't changed its status

UNIT TEST
added a unit test to make sure the VM is kept in the map
although the call to GetVmStats didn't made it

Bug-Url: https://bugzilla.redhat.com/1072282
Change-Id: I415804d9fa1f1288423241b8547f4cab5540914a
Signed-off-by: Roy Golan <rgolan@redhat.com>