core: disable command's logging on runningFailed
After calling RunVmCommandBase#runningFailed the automatic logging
should be disabled since the 'log()' method is called directly by this
method. A call to 'setCommandShouldBeLogged(false)' was added inside
RunVmCommandBase#runningFailed.

It is safe since this method is called on rerun flows, where the
automatic logging is not invoked anyway, and from one place as part of
the execution phase where the automatic logging is already disabled.

Change-Id: Ied9130d2e8dacb94924da220e8ba1a1f1204e681
Signed-off-by: Arik Hadas <ahadas@redhat.com>