core, engine, webadmin: Shutdown/power off operations on a VM would ask for an optional reason
When a VM is shutdown/powered off from the mananger there
should be an optional "reason" field where the administrator
can write a free-form explanation for the operation, which
would be bought to attention in the logs and when/if someone
attempts to power on.

Adds a Reason column to VMs tab which is populated if a
reason was specified during shutdown/power off operations.

The reason is cleared when the VM is successfully started again

The feature can be enabled/disabled at the cluster level.

Change-Id: I17645e5bc97a1a4d460956ec45f88524465dfd7b
Bug-Url: https://bugzilla.redhat.com/1065753
Signed-off-by: Ravi Nori <rnori@redhat.com>