core: TagsDirector: stop using compat's Regex
org.ovirt.engine.core.compat.Regex is deprecated, and should only be
used for GWT-compiled classes.

This patch removes the call to Regex.IsMatch(String, String), and
replaces it with calls to equivelant calls to
java.util.regexp.Pattern.

Note: This patch does nothing but unwrap the Regex call. It's possible
there's a better way of achieving the same functionality, but that
should be left to a different patch.

Change-Id: I98c1bc6204a7bb8642e3f215b1547260f36b0152
Signed-off-by: Allon Mureinik <amureini@redhat.com>