webadmin: Only set cluster once when opening VM dialog
Previously the cluster could potentially be called twice in the
initialization of the dialog: once UiCommonEditorVisitor set the
selected item to be the first in the collection, then if the VM
belonged to another cluster the selection would change. This would
cause issues due to asynchronous tasks, triggered by the
selectedItemChangedEvent, returning in reversed order.

This is fixed by using the new overload of ListModel.setItems(), to
make sure that the selected cluster is properly selected the first
time and so won't trigger dual sets of backend queries.

Change-Id: I5b4dd3328f207dadb73ee93d27f084249f3f9704
Bug-Url: https://bugzilla.redhat.com/1070835
Signed-off-by: Lior Vernia <lvernia@redhat.com>