core: move constant strings in ovf to OvfProperties
Most of the Strings which are part of the OVF files appear explicitly
in both the reader and writer classes. It is wrong because they have to
be the same in both places. So in this patch I move many Strings like
that to the OvfProperties class as constants and the OVF readers and
writers now reference those constants.

Change-Id: Idc40102b107d510e04f10f45f0b7e4838fd4c706
Signed-off-by: Arik Hadas <ahadas@redhat.com>