core: Introducing OvfStoreOnAnyDomain feature
this patch intrdouces the OvfOnAnyDomain feature -
For any domain in supported pool the engine will attempt to store OVFs
of VMs/Templates that have image disks on that domain.
As soon as there will be such vms/templates, the engine will attempt to
create OVF disk on that domain. If the disk won't be created
succesfully, it will remain in ILLEGAL status untill it will be removed
by the user.

As soon as the OVF disk is created succesfully for the domain, an in
memory tar with all the related OVF data will be created and uploaded to
the ovf store disk on the domain.

Currently most of the operation are blocked for ovf disks, they will be
enabled in further patches.

Change-Id: I78803db10947a6ee477a493b530319654b1c045d
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1032686
Signed-off-by: Liron Aravot <laravot@redhat.com>