core: performance improvement
* initialized ArrayList to avoid copying arrays (tradeof is
bigger memory footprint during initialization)

* changed transforming macAddress stored in 'long' to
String; loop is used instead of regex.

---
this changes cuts CPU time approx to 50% of original.
---
Note: code can be suboptimal when looking for LOT OF macs but range
offers to little of them: in that case unneccesarily big
array alocated and then trown away for no reason. But this should
not be an issue.

Change-Id: I09f87ebd1ea17d09a974ce6ca3e26bd2454a72fc
Bug-Url: https://bugzilla.redhat.com/1063064
Signed-off-by: Martin Mucha <mmucha@redhat.com>