tools : engine-config is over writing the previous values
'engine-config -s' is over writing the existing values
with the new ones from command line.

A new option to merge the new values provided with the
values in the database has been added.

'engine-config -m' or 'engine-config --merge'

This concatenates the new value with the exisiting value
in the database.

Change-Id: I5c70539f47c509e3b8c23b1aa3de41bead36c1b4
Bug-Url: https://bugzilla.redhat.com/1069768
Signed-off-by: Ravi Nori <rnori@redhat.com>